+++ waves/builders.py
@@ -229,11 +229,10 @@ def _construct_post_action_list(post_action):
     return new_actions
 
 
-def _abaqus_journal_emitter(target, source, env):
-    """Appends the abaqus_journal builder target list with the builder managed targets
+def _first_target_emitter(target, source, env, suffixes=[_stdout_extension]):
+    """Appends the target list with the builder managed targets
 
-    Appends ``target[0]``.stdout and ``target[0]``.abaqus_v6.env to the ``target`` list. The abaqus_journal Builder
-    requires at least one target.
+    Appends ``target[0]``.stdout to the ``target`` list. The associated Builder requires at least one target.
 
     The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
     a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
@@ -252,11 +251,32 @@ def _abaqus_journal_emitter(target, source, env):
         build_subdirectory = first_target.parents[0]
     except IndexError as err:
         build_subdirectory = pathlib.Path(".")
-    suffixes = [_stdout_extension, _abaqus_environment_extension]
     for suffix in suffixes:
         emitter_target = build_subdirectory / first_target.with_suffix(suffix).name
         target.append(str(emitter_target))
     return target, source
+
+
+def _abaqus_journal_emitter(target, source, env):
+    """Appends the abaqus_journal builder target list with the builder managed targets
+
+    Appends ``target[0]``.stdout and ``target[0]``.abaqus_v6.env to the ``target`` list. The abaqus_journal Builder
+    requires at least one target.
+
+    The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
+    a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
+    subdirectory, e.g. ``parameter_set1/target.ext``. When in doubt, provide the expected STDOUT redirected file as a
+    target, e.g. ``target[0].stdout``.
+
+    :param list target: The target file list of strings
+    :param list source: The source file list of SCons.Node.FS.File objects
+    :param SCons.Script.SConscript.SConsEnvironment env: The builder's SCons construction environment object
+
+    :return: target, source
+    :rtype: tuple with two lists
+    """
+    suffixes = [_stdout_extension, _abaqus_environment_extension]
+    return _first_target_emitter(target, source, env, suffixes=suffixes)
 
 
 def abaqus_journal(abaqus_program="abaqus", post_action=[]):
@@ -449,35 +469,6 @@ def copy_substitute(source_list, substitution_dictionary={}, env=SCons.Environme
     return target_list
 
 
-def _python_script_emitter(target, source, env):
-    """Appends the python_script builder target list with the builder managed targets
-
-    Appends ``target[0]``.stdout to the ``target`` list. The python_script Builder requires at least one target.
-
-    The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
-    a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
-    subdirectory, e.g. ``parameter_set1/target.ext``. When in doubt, provide the expected STDOUT redirected file as a
-    target, e.g. ``target[0].stdout``.
-
-    :param list target: The target file list of strings
-    :param list source: The source file list of SCons.Node.FS.File objects
-    :param SCons.Script.SConscript.SConsEnvironment env: The builder's SCons construction environment object
-
-    :return: target, source
-    :rtype: tuple with two lists
-    """
-    first_target = pathlib.Path(str(target[0]))
-    try:
-        build_subdirectory = first_target.parents[0]
-    except IndexError as err:
-        build_subdirectory = pathlib.Path(".")
-    suffixes = [_stdout_extension]
-    for suffix in suffixes:
-        emitter_target = build_subdirectory / first_target.with_suffix(suffix).name
-        target.append(str(emitter_target))
-    return target, source
-
-
 def python_script(post_action=[]):
     """Python script SCons builder
 
@@ -521,8 +512,50 @@ def python_script(post_action=[]):
     action.extend(_construct_post_action_list(post_action))
     python_builder = SCons.Builder.Builder(
         action=action,
-        emitter=_python_script_emitter)
+        emitter=_first_target_emitter)
     return python_builder
+
+
+def matlab_script(matlab_program="matlab", post_action=[], symlink=False):
+    """Matlab script SCons builder
+
+    .. warning::
+
+       Experimental implementation is subject to change
+
+    This builder requires that the Matlab script to execute is provided by absolute path and the first source in the
+    list. The builder returned by this function accepts all SCons Builder arguments and adds the ``script_options`` and
+    ``matlab_options`` string arguments.
+
+    First the Matlab script is copied to the first target file's build directory and then the script is executed
+    locally.
+
+    At least one target must be specified. The first target determines the working directory for the builder's action,
+    as shown in the action code snippet below. The action changes the working directory to the first target's parent
+    directory prior to executing the python script.
+
+    The Builder emitter will append the builder managed targets automatically. Appends ``target[0]``.stdout to the
+    ``target`` list.
+
+    The emitter will assume all emitted targets build in the current build directory. If the target(s) must be built in
+    a build subdirectory, e.g. in a parameterized target build, then the first target must be provided with the build
+    subdirectory, e.g. ``parameter_set1/my_target.ext``. When in doubt, provide the expected STDOUT redirected file as a
+    target, e.g. ``my_target.stdout``.
+
+    .. code-block::
+       :caption: Matlab script builder action
+
+       Copy("${TARGET.dir.abspath}", ${SOURCE.abspath}, False)
+       cd ${TARGET.dir.abspath} && {matlab_program} ${matlab_options} -batch "${SOURCE.filebase}(${script_options})" > ${TARGET.filebase}.stdout 2>&1
+    """
+    action = [SCons.Defaults.Copy("${TARGET.dir.abspath}", "${SOURCE.abspath}", symlink),
+              f"{_cd_action_prefix} {matlab_program} ${{matlab_options}} -batch \"${{SOURCE.filebase}}(" \
+                f"${{script_options}})\" > ${{TARGET.filebase}}{_stdout_extension} 2>&1"]
+    action.extend(_construct_post_action_list(post_action))
+    matlab_builder = SCons.Builder.Builder(
+        action=action,
+        emitter=_first_target_emitter)
+    return matlab_builder
 
 
 def conda_environment():
@@ -670,30 +703,6 @@ def _build_odb_extract(target, source, env):
     return None
 
 
-def _sbatch_emitter(target, source, env):
-    """Appends the sbatch builder target list with the builder managed targets
-
-    Appends ``target[0]``.stdout to the ``target`` list. The sbatch Builder requires at least one target.
-
-    :param list target: The target file list of strings
-    :param list source: The source file list of SCons.Node.FS.File objects
-    :param SCons.Script.SConscript.SConsEnvironment env: The builder's SCons construction environment object
-
-    :return: target, source
-    :rtype: tuple with two lists
-    """
-    first_target = pathlib.Path(str(target[0]))
-    try:
-        build_subdirectory = first_target.parents[0]
-    except IndexError as err:
-        build_subdirectory = pathlib.Path(".")
-    suffixes = [_stdout_extension]
-    for suffix in suffixes:
-        emitter_target = build_subdirectory / first_target.with_suffix(suffix).name
-        target.append(str(emitter_target))
-    return target, source
-
-
 def sbatch(sbatch_program="sbatch", post_action=[]):
     """SLURM sbatch SCons builder
 
@@ -732,5 +741,5 @@ def sbatch(sbatch_program="sbatch", post_action=[]):
     action.extend(_construct_post_action_list(post_action))
     sbatch_builder = SCons.Builder.Builder(
         action=action,
-        emitter=_sbatch_emitter)
+        emitter=_first_target_emitter)
     return sbatch_builder