+++ waves/tests/test_salib_sampler.py
@@ -71,6 +71,17 @@ class TestSALibSampler:
             },
             pytest.raises(ValueError)
         ),
+        "morris: one parameter": (
+            "morris",
+            {"N": 4,
+             "problem": {
+                 "num_vars": 1,
+                 "names": ["parameter_1",],
+                 "bounds": [[-1, 1]]
+             }
+            },
+            pytest.raises(ValueError)
+        ),
         "missing N": (
             "latin",
             {"problem": {"num_vars": 4, "names": ["p1"], "bounds": [[-1, 1]]}},
@@ -152,6 +163,14 @@ class TestSALibSampler:
                          "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
             },
             {"seed": 42},
+        ),
+        "good schema 65x3": (
+            {"N": 65,
+             "problem": {"num_vars": 3,
+                         "names": ["parameter_1", "parameter_2", "parameter_3"],
+                         "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
+            },
+            {"seed": 42},
         )
     }
 
@@ -161,7 +180,23 @@ class TestSALibSampler:
             number_of_simulations = N * (num_vars + 2)
         elif sampler == "sobol":
             number_of_simulations = N * (2 * num_vars + 2)
+        elif sampler == "fast_sampler":
+            number_of_simulations = N * num_vars
+        elif sampler == "finite_diff":
+            number_of_simulations = N * (num_vars + 1)
+        elif sampler == "morris":
+            # Default interface settings
+            number_of_simulations = int((num_vars + 1) * N)
         return [f"parameter_set{num}" for num in range(number_of_simulations)]
+
+    def _big_enough(self, sampler, N, num_vars):
+        if sampler == "sobol" and num_vars < 2:
+            return False
+        elif sampler == "fast_sampler" and N < 64:
+            return False
+        elif sampler == "morris" and num_vars < 2:
+            return False
+        return True
 
     @pytest.mark.unittest
     @pytest.mark.parametrize("parameter_schema, kwargs",
@@ -170,20 +205,24 @@ class TestSALibSampler:
     def test_generate(self, parameter_schema, kwargs):
         for sampler in _supported_salib_samplers:
             # TODO: find a better way to separate the sampler types and their test parameterization
-            if sampler == "sobol" and parameter_schema["problem"]["num_vars"] < 2:
+            if not self._big_enough(sampler, parameter_schema["N"], parameter_schema["problem"]["num_vars"]):
                 return
             # Unit tests
             TestGenerate = SALibSampler(sampler, parameter_schema, **kwargs)
             samples_array = TestGenerate._samples
+            assert samples_array.shape[1] == parameter_schema["problem"]["num_vars"]
             # Verify that the parameter set name creation method was called
-            expected_set_names = self._expected_set_names(sampler, parameter_schema["N"], parameter_schema["problem"]["num_vars"])
-            assert list(TestGenerate._parameter_set_names.values()) == expected_set_names
-            # Check that the parameter set names are correctly populated in the parameter study Xarray Dataset
-            parameter_set_names = list(TestGenerate.parameter_study[_set_coordinate_key])
-            assert numpy.all(parameter_set_names == expected_set_names)
+            # Morris produces inconsistent set counts depending on seed. Rely on the variable count shape check above.
+            if not sampler == "morris":
+                expected_set_names = self._expected_set_names(sampler, parameter_schema["N"], parameter_schema["problem"]["num_vars"])
+                assert samples_array.shape[0] == len(expected_set_names)
+                assert list(TestGenerate._parameter_set_names.values()) == expected_set_names
+                # Check that the parameter set names are correctly populated in the parameter study Xarray Dataset
+                parameter_set_names = list(TestGenerate.parameter_study[_set_coordinate_key])
+                assert numpy.all(parameter_set_names == expected_set_names)
 
     merge_test = {
-        "new sets, 2 params": (
+        "new sets, 5(8)x2": (
             {"N": 5,
              "problem": {"num_vars": 2,
                          "names": ["parameter_1", "parameter_2"],
@@ -196,7 +235,7 @@ class TestSALibSampler:
             },
             {"seed": 42},
         ),
-        "new sets, 3 params": (
+        "new sets, 5(8)x3": (
             {"N": 5,
              "problem": {"num_vars": 3,
                          "names": ["parameter_1", "parameter_2", "parameter_3"],
@@ -209,7 +248,7 @@ class TestSALibSampler:
             },
             {"seed": 42},
         ),
-        "unchanged sets, 2 param": (
+        "unchanged sets, 5x2": (
             {"N": 5,
              "problem": {"num_vars": 2,
                          "names": ["parameter_1", "parameter_2"],
@@ -222,13 +261,39 @@ class TestSALibSampler:
             },
             {"seed": 42},
         ),
-        "unchanged sets, 3 param": (
+        "unchanged sets, 5x3": (
             {"N": 5,
              "problem": {"num_vars": 3,
                          "names": ["parameter_1", "parameter_2", "parameter_3"],
                          "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
             },
             {"N": 5,
+             "problem": {"num_vars": 3,
+                         "names": ["parameter_1", "parameter_2", "parameter_3"],
+                         "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
+            },
+            {"seed": 42},
+        ),
+        "changed sets, 65(70)x3": (
+            {"N": 65,
+             "problem": {"num_vars": 3,
+                         "names": ["parameter_1", "parameter_2", "parameter_3"],
+                         "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
+            },
+            {"N": 70,
+             "problem": {"num_vars": 3,
+                         "names": ["parameter_1", "parameter_2", "parameter_3"],
+                         "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
+            },
+            {"seed": 42},
+        ),
+        "unchanged sets, 65x3": (
+            {"N": 65,
+             "problem": {"num_vars": 3,
+                         "names": ["parameter_1", "parameter_2", "parameter_3"],
+                         "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
+            },
+            {"N": 65,
              "problem": {"num_vars": 3,
                          "names": ["parameter_1", "parameter_2", "parameter_3"],
                          "bounds": [[-1, 1], [-2, 2], [-3, 3]]},
@@ -243,6 +308,10 @@ class TestSALibSampler:
                              ids=merge_test.keys())
     def test_merge(self, first_schema, second_schema, kwargs):
         for sampler in _supported_salib_samplers:
+            # TODO: find a better way to separate the sampler types and their test parameterization
+            if not self._big_enough(sampler, first_schema["N"], first_schema["problem"]["num_vars"]):
+                return
+            # Unit tests
             TestMerge1 = SALibSampler(sampler, first_schema, **kwargs)
             with patch('xarray.open_dataset', return_value=TestMerge1.parameter_study):
                 TestMerge2 = SALibSampler(sampler, second_schema, previous_parameter_study='dummy_string', **kwargs)