+++ docs/changelog.rst
@@ -5,11 +5,35 @@ Changelog
 #########
 
 ******************
-0.6.2 (unreleased)
+0.6.3 (unreleased)
 ******************
 
 ******************
-0.6.1 (2022-01-06)
+0.6.2 (2023-01-13)
+******************
+
+New Features
+============
+- Add SALib ``fast_sampler`` to the list of tested samplers for parameter generation (:merge:`444`). By `Kyle
+  Brindley`_.
+- Add SALib ``finite_diff`` to the list of tested samplers (:merge:`447`). By `Kyle Brindley`_.
+- Add SALib ``morris`` to the list of tested samplers (:issue:`386`, :merge:`443`). By `Kyle Brindley`_.
+- Add an experimental draft builder for Matlab scripts (:issue:`388`, :merge:`449`). By `Kyle Brindley`_.
+
+Documentation
+=============
+- Adjust PDF documentation build's font size of code-blocks to fit 120 character width files (:merge:`445` :merge:`446`). By `Kyle
+  Brindley`_.
+- Remove unnecessary nested f-string and SCons variable replacement syntax from post-processing tasks in core tutorials
+  (:issue:`377`, :merge:`448`). By `Kyle Brindley`_.
+
+Internal Changes
+================
+- Reduce builder emitter code duplication with a common "first target" emitter (:issue:`388`, :merge:`449`,
+  :merge:`450`). By `Kyle Brindley`_.
+
+******************
+0.6.1 (2023-01-06)
 ******************
 
 New Features
@@ -25,7 +49,7 @@ Internal Changes
   with a deprecation warning (:issue:`381`, :merge:`440`). By `Kyle Brindley`_.
 
 *******************
-0.5.11 (2022-01-05)
+0.5.11 (2023-01-05)
 *******************
 
 New Features